Create message_list
